🚚 moved all issues components inside of issues folder
