+++ src/app/(components)/issues/clear-search-button.tsx
